babelia/counter.scm: pack prefix.
okvs abstraction prefix must always be a list of scheme objects that
can be packed by okvs' pack procedure, because from the user
perspective, it is easier to reason about okvs as scheme object
key-value space instead of bytevector key-value space.

At least, that is the rationale for why nstore's prefix is a list
instead of a bytevector. At least, to be consistent with nstore
counter's prefix must be a list.