babelia/okvs/mapping/tests: change prefix
Since the very previous commit, it is possible to use a list as a
prefix without having to keep track of the length of the prefix et al.
as currently done in nstore.

This is a compromise between ease-of-use and space performance.

This is different from what FoundationDB does. Unlike FDB, okvs pack
does not expose a bytevector prefix argument, that allows to have zero
space overhead when using bytes as subspace prefix because
fdb.tuple.pack will strip the prefix bytes passed as an argument
before parsing the bytes. Eventually, that will lead the user to
juggle between bytes (the various subspace prefixes) and higher level
data types (integer, float, string, symbol, and list). That is, it
make things more complicated.

The space overhead that is prescribed by requiring the prefix to be
packed by the same packing procedure is a single byte, plus one byte
overhead per null bytes in bytevectors or nested lists.  That is, one
or more bytes of overhead per key-value pair.

The choosen appraoch has the advantage of making it possible to unpack
the whole database and discover the schema. That is, to be able to
unpack a stored bytevector, it is not required to know or guess the
subspace bytes prefix.

It must be noted also, that since prefix can be a list, it is easy, to
compose abstractions and keep the database "readable". For instance,
in case the mapping is nested inside another subspace, the prefix of
the higher subspace, must be prefix of the nested mapping prefix:

  (make-mapping engine (append super-prefix 'mapping))

In FoundationDB approach, to do something similar would require to append
bytevectors:

  (make-mapping engine (bytevector-append super-prefix #vu8(42)))

Hence, all the keys, in the defined mapping will have as prefix:

  (bytevector-append super-prefix #vu8(42))

That is much less readable than:

  (list (list 'some 'super 'prefix 'mapping) 'my-key)

Where 'my-key is a key in the defined mapping. Note, that 'my-key is
simply the cadr of the whole list. No need to keep track of prefix
length, at this level of abstraction, to extract the interesting
stuff.

One can argue that this approach leads to one bytes of overhead, and
certainly more when the prefix is a bytevector with null bytes (see
the algorithm of bytevectors packing). Indeed, when performance is
required, one must rely on bytevector as prefix. In the current
approach, it is also better to avoid null bytes in bytevector
prefixes.

Like noted earlier, the approach that is taken will optimize for ease
of use and hopefully prescribes a smooth learning curve.